The latest version of the Task Skill tutorial file.
The change is to demonstrate the colimit for assembling three different
schemas sS, sT, sSnT (and associated instances) each with one table. I
then used an Uber Flower to wire up the tables with foreign keys the
result of which is a schema diagram that looks like the UMP for products
whith a particular intersection table describing which skills are needed
to perform a task.